// .PRD/erp-refactor/v3/types/purchase-receipt.ts
export const PurchaseReceiptSchema = z.object({
  id: z.string(),
  owner: z.string().optional(),
  creation: z.string().optional(),
  modified: z.string().optional(),
  modified_by: z.string().optional(),
  docstatus: z.number().default(0),
  title: z.string().optional().default('{supplier_name}'),
  naming_series: z.enum(['MAT-PRE-.YYYY.-', 'MAT-PR-RET-.YYYY.-']),
  supplier: z.string(),
  supplier_name: z.string().optional(),
  supplier_delivery_note: z.string().optional(),
  subcontracting_receipt: z.string().optional(),
  posting_date: z.string().default('Today'),
  posting_time: z.string().default('Now'),
  set_posting_time: z.boolean().optional().default(false),
  company: z.string(),
  apply_putaway_rule: z.boolean().optional().default(false),
  is_return: z.boolean().optional().default(false),
  return_against: z.string().optional(),
  cost_center: z.string().optional(),
  project: z.string().optional(),
  currency: z.string(),
  conversion_rate: z.number(),
  buying_price_list: z.string().optional(),
  price_list_currency: z.string().optional(),
  plc_conversion_rate: z.number().optional(),
  ignore_pricing_rule: z.boolean().optional().default(false),
  scan_barcode: z.string().optional(),
  last_scanned_warehouse: z.string().optional(),
  set_warehouse: z.string().optional(),
  set_from_warehouse: z.string().optional(),
  rejected_warehouse: z.string().optional(),
  is_subcontracted: z.boolean().optional().default(false),
  supplier_warehouse: z.string().optional(),
  items: z.array(z.unknown()),
  total_qty: z.number().optional(),
  total_net_weight: z.number().optional(),
  base_total: z.number().optional(),
  base_net_total: z.number(),
  total: z.number().optional(),
  net_total: z.number().optional(),
  tax_category: z.string().optional(),
  taxes_and_charges: z.string().optional(),
  shipping_rule: z.string().optional(),
  incoterm: z.string().optional(),
  named_place: z.string().optional(),
  taxes: z.array(z.unknown()).optional(),
  base_taxes_and_charges_added: z.number().optional(),
  base_taxes_and_charges_deducted: z.number().optional(),
  base_total_taxes_and_charges: z.number().optional(),
  taxes_and_charges_added: z.number().optional(),
  taxes_and_charges_deducted: z.number().optional(),
  total_taxes_and_charges: z.number().optional(),
  grand_total: z.number().optional(),
  disable_rounded_total: z.boolean().optional().default(false),
  rounding_adjustment: z.number().optional(),
  in_words: z.string().max(240).optional(),
  rounded_total: z.number().optional(),
  base_grand_total: z.number().optional(),
  base_rounding_adjustment: z.number().optional(),
  base_in_words: z.string().max(240).optional(),
  base_rounded_total: z.number().optional(),
  apply_discount_on: z.enum(['Grand Total', 'Net Total']).optional().default('Grand Total'),
  base_discount_amount: z.number().optional(),
  additional_discount_percentage: z.number().optional(),
  discount_amount: z.number().optional(),
  other_charges_calculation: z.string().optional(),
  item_wise_tax_details: z.array(z.unknown()).optional(),
  pricing_rules: z.array(z.unknown()).optional(),
  supplied_items: z.array(z.unknown()).optional(),
  supplier_address: z.string().optional(),
  address_display: z.string().optional(),
  contact_person: z.string().optional(),
  contact_display: z.string().optional(),
  contact_mobile: z.string().optional(),
  contact_email: z.string().optional(),
  dispatch_address: z.string().optional(),
  dispatch_address_display: z.string().optional(),
  shipping_address: z.string().optional(),
  shipping_address_display: z.string().optional(),
  billing_address: z.string().optional(),
  billing_address_display: z.string().optional(),
  tc_name: z.string().optional(),
  terms: z.string().optional(),
  status: z.enum(['Draft', 'Partly Billed', 'To Bill', 'Completed', 'Return', 'Return Issued', 'Cancelled', 'Closed']).default('Draft'),
  per_billed: z.number().optional(),
  per_returned: z.number().optional(),
  auto_repeat: z.string().optional(),
  letter_head: z.string().optional(),
  group_same_items: z.boolean().optional().default(false),
  select_print_heading: z.string().optional(),
  language: z.string().optional(),
  transporter_name: z.string().optional(),
  lr_no: z.string().optional(),
  lr_date: z.string().optional(),
  instructions: z.string().optional(),
  is_internal_supplier: z.boolean().optional().default(false),
  represents_company: z.string().optional(),
  inter_company_reference: z.string().optional(),
  remarks: z.string().optional(),
  range: z.string().optional(),
  amended_from: z.string().optional(),
  is_old_subcontracting_flow: z.boolean().optional().default(false),
  other_details: z.string().optional(),
});

export type PurchaseReceipt = z.infer<typeof PurchaseReceiptSchema>;

export const PurchaseReceiptInsertSchema = PurchaseReceiptSchema.omit({ id: true, owner: true, creation: true, modified: true, modified_by: true });